Add new unit tests for gated delta net padding and parallel configurations
This commit introduces additional unit tests for the gated delta net, specifically focusing on the correctness of padding and parallel execution. New test cases have been added to the CI configuration to ensure comprehensive coverage for these scenarios.
